MetaDataMDX __getitem__ thread saftey
Refactor the __getitem__ method on the MetaDataMDX class so that
it is thread safe. Different threads may be simultaneously passing
through the method but take different paths through the if, elif, else
and so have different states for the local variable entity. As such
the syntax self.entity[item] is ambiguous and can be evaluated different
ways depending on the thread state, resulting in a race condition.